bot/requirements.txt
aiogram==3.*
asyncpg==0.*
python-dotenv==1.*